+++ src/EntityAbstractBuilder.php
@@ -658,35 +658,35 @@ MBODY;
         // ===============================================
 
         // METHOD:getArrayCopy
-        $method = $this->buildMethod('getArrayCopy');
-        $objectParam = new ParameterGenerator('object', '?object');
-        $objectParam->setDefaultValue(null);
-        $method->setParameter($objectParam);
-        $method->setParameter(new ParameterGenerator('ignorePrimaryColumn', 'bool', false));
-        $method->setReturnType('array');
-        $mjson = [];
-        foreach ($this->getNames()->properties as $name => $prop) if ($prop['type'] == 'array') $mjson[] = "\$data['{$name}'] = Json::encode(\$this->data['{$name}']);";
-        $mjson = implode("\n", $mjson);
+//         $method = $this->buildMethod('getArrayCopy');
+//         $objectParam = new ParameterGenerator('object', '?object');
+//         $objectParam->setDefaultValue(null);
+//         $method->setParameter($objectParam);
+//         $method->setParameter(new ParameterGenerator('ignorePrimaryColumn', 'bool', false));
+//         $method->setReturnType('array');
+//         $mjson = [];
+//         foreach ($this->getNames()->properties as $name => $prop) if ($prop['type'] == 'array') $mjson[] = "\$data['{$name}'] = Json::encode(\$this->data['{$name}']);";
+//         $mjson = implode("\n", $mjson);
 
-        $body = <<<MBODY
-\$data = array_merge([], \$this->data);
-{$mjson}
-if (\$ignorePrimaryColumn) foreach (\$this->primaryKeyColumn as \$column) unset(\$data[\$column]);
-return \$data;
-MBODY;
-        $method->setBody($body);
-        $docBlock = new DocBlockGenerator();
-        $docBlock->setTag(new ParamTag('object', [
-            'datatype' => '?object'
-        ]));
-        $docBlock->setTag(new ParamTag('ignorePrimaryColumn', [
-            'datatype' => 'bool'
-        ]));
-        $docBlock->setTag(new ReturnTag([
-            'datatype' => 'Array'
-        ]));
-        $docBlock->setShortDescription("Array copy of object");
-        $method->setDocBlock($docBlock);
-        $this->_class->addMethodFromGenerator($method);
+//         $body = <<<MBODY
+// \$data = array_merge([], \$this->data);
+// {$mjson}
+// if (\$ignorePrimaryColumn) foreach (\$this->primaryKeyColumn as \$column) unset(\$data[\$column]);
+// return \$data;
+// MBODY;
+//         $method->setBody($body);
+//         $docBlock = new DocBlockGenerator();
+//         $docBlock->setTag(new ParamTag('object', [
+//             'datatype' => '?object'
+//         ]));
+//         $docBlock->setTag(new ParamTag('ignorePrimaryColumn', [
+//             'datatype' => 'bool'
+//         ]));
+//         $docBlock->setTag(new ReturnTag([
+//             'datatype' => 'Array'
+//         ]));
+//         $docBlock->setShortDescription("Array copy of object");
+//         $method->setDocBlock($docBlock);
+//         $this->_class->addMethodFromGenerator($method);
     }
 }